+++ tests/unitTests/material_model_tests/test_material_mooney_rivlin.h
@@ -1,0 +1,108 @@
+/* Copyright (c) Stanford University, The Regents of the University of California, and others.
+ *
+ * All Rights Reserved.
+ *
+ * See Copyright-SimVascular.txt for additional details.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject
+ * to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+ * OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#ifndef TEST_MATERIAL_MOONEY_RIVLIN_H
+#define TEST_MATERIAL_MOONEY_RIVLIN_H
+
+#include "test_material_common.h"
+
+
+
+// Class to contain Mooney-Rivlin material parameters
+class MooneyRivlinParams : public MatParams {
+public:
+    double C01;
+    double C10;
+
+    // Default constructor
+    MooneyRivlinParams() : C01(0.0), C10(0.0) {}
+
+    // Constructor with parameters
+    MooneyRivlinParams(double c01, double c10) : C01(c01), C10(c10) {}
+
+};
+
+
+/**
+ * @brief Class for testing the Mooney-Rivlin material model.
+ *
+ * This class provides methods to set up and test the Mooney-Rivlin material model, including 
+ * computing the strain energy and printing material parameters.
+ */
+class TestMooneyRivlin : public TestMaterialModel {
+public:
+
+    /**
+     * @brief Parameters for the Mooney-Rivlin material model.
+     */
+    MooneyRivlinParams params;
+
+    /**
+     * @brief Constructor for the TestMooneyRivlin class.
+     *
+     * Initializes the Mooney-Rivlin material parameters for svMultiPhysics.
+     *
+     * @param[in] params_ Parameters for the Mooney-Rivlin material model.
+     */
+    TestMooneyRivlin(const MooneyRivlinParams &params_) : TestMaterialModel( consts::ConstitutiveModelType::stIso_MR, consts::ConstitutiveModelType::stVol_ST91),
+        params(params_) 
+        {
+        // Set Mooney-Rivlin material parameters for svMultiPhysics
+        auto &dmn = com_mod.mockEq.mockDmn;
+        dmn.stM.C01 = params.C01;
+        dmn.stM.C10 = params.C10;
+        dmn.stM.Kpen = 0.0;         // Zero volumetric penalty parameter
+    }
+
+    /**
+     * @brief Prints the Mooney-Rivlin material parameters.
+     */
+    void printMaterialParameters() {
+        std::cout << "C01 = " << params.C01 << ", C10 = " << params.C10 << std::endl;
+    }
+
+    /**
+     * @brief Computes the strain energy for the Mooney-Rivlin material model.
+     *
+     * @param[in] F Deformation gradient.
+     * @return Strain energy density for the Mooney-Rivlin material model.
+     */
+    double computeStrainEnergy(const Array<double> &F) {
+        // Compute solid mechanics terms
+        solidMechanicsTerms smTerms = calcSolidMechanicsTerms(F);
+
+        // Strain energy density for Mooney-Rivlin material model
+        // Psi_iso = C10 * (Ib1 - 3) + C01 * (Ib2 - 3)
+        double Psi_iso = params.C10 * (smTerms.Ib1 - 3.) + params.C01 * (smTerms.Ib2 - 3.);
+
+        return Psi_iso;
+    }
+};
+#endif // TEST_MATERIAL_MOONEY_RIVLIN_H
