+++ tests/unitTests/material_model_tests/test_material_neohookean.h
@@ -1,0 +1,106 @@
+/* Copyright (c) Stanford University, The Regents of the University of California, and others.
+ *
+ * All Rights Reserved.
+ *
+ * See Copyright-SimVascular.txt for additional details.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject
+ * to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+ * OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#ifndef TEST_MATERIAL_NEOHOOKEAN_H
+#define TEST_MATERIAL_NEOHOOKEAN_H
+
+#include "test_material_common.h"
+
+
+
+
+// Class to contain Neo-Hookean material parameters
+class NeoHookeanParams : public MatParams {
+public:
+    double C10;
+
+    // Default constructor
+    NeoHookeanParams() : C10(0.0) {}
+
+    // Constructor with parameters
+    NeoHookeanParams(double c10) : C10(c10) {}
+
+};
+
+/**
+ * @brief Class for testing the Neo-Hookean material model.
+ *
+ * This class provides methods to set up and test the Neo-Hookean material model, including 
+ * computing the strain energy and printing material parameters.
+ */
+class TestNeoHookean : public TestMaterialModel {
+public:
+
+    /**
+     * @brief Parameters for the Neo-Hookean material model.
+     */
+    NeoHookeanParams params;
+
+    /**
+     * @brief Constructor for the TestNeoHookean class.
+     *
+     * Initializes the Neo-Hookean material parameters for svMultiPhysics.
+     *
+     * @param[in] params_ Parameters for the Neo-Hookean material model.
+     */
+    TestNeoHookean(const NeoHookeanParams &params_) : TestMaterialModel( consts::ConstitutiveModelType::stIso_nHook, consts::ConstitutiveModelType::stVol_ST91),
+        params(params_) 
+        {
+        // Set Neo-Hookean material parameters for svMultiPhysics
+        auto &dmn = com_mod.mockEq.mockDmn;
+        dmn.stM.C10 = params.C10;
+        dmn.stM.Kpen = 0.0;         // Zero volumetric penalty parameter
+    }
+
+    /**
+     * @brief Prints the Neo-Hookean material parameters.
+     */
+    void printMaterialParameters() {
+        std::cout << "C10 = " << params.C10 << std::endl;
+    }
+
+    /**
+     * @brief Computes the strain energy for the Neo-Hookean material model.
+     *
+     * @param[in] F Deformation gradient.
+     * @return Strain energy density for the Neo-Hookean material model.
+     */
+    double computeStrainEnergy(const Array<double> &F) {
+        // Compute solid mechanics terms
+        solidMechanicsTerms smTerms = calcSolidMechanicsTerms(F);
+
+        // Strain energy density for Neo-Hookean material model
+        // Psi_iso = C10 * (Ib1 - 3)
+        double Psi_iso = params.C10 * (smTerms.Ib1 - 3.);
+
+        return Psi_iso;
+    }
+};
+#endif
