+++ tests/unitTests/test_common.h
@@ -1,0 +1,99 @@
+/* Copyright (c) Stanford University, The Regents of the University of California, and others.
+ *
+ * All Rights Reserved.
+ *
+ * See Copyright-SimVascular.txt for additional details.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject
+ * to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+ * OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+#ifndef TEST_COMMON_H
+#define TEST_COMMON_H
+
+#include <stdlib.h>
+#include <iostream>
+#include <random>
+#include <chrono>
+#include "CepMod.h"
+#include "ComMod.h"
+#include "gtest/gtest.h"
+
+
+// --------------------------------------------------------------
+// -------------------- Mock svMultiPhysics object -------------------
+// --------------------------------------------------------------
+
+
+class MockCepMod : public CepMod {
+public:
+    MockCepMod() {
+        // initialize if needed 
+    }
+    // Mock methods if needed
+};
+class MockdmnType : public dmnType {
+public:
+    MockdmnType() {
+        // initialize if needed 
+    }
+    // MockstModelType mockStM;
+    // Mock methods if needed
+};
+class MockmshType : public mshType {
+public:
+    MockmshType() {
+        // initialize if needed 
+    }
+    // Mock methods if needed
+};
+class MockeqType : public eqType {
+public:
+    MockeqType() {
+        // initialize if needed 
+    }
+    MockdmnType mockDmn;
+    // Mock methods if needed
+};
+class MockComMod : public ComMod {
+public:
+    MockComMod() {
+        // initialize if needed 
+        nsd = 3;
+    }
+    MockeqType mockEq;
+    MockmshType mockMsh;
+    // Mock methods if needed
+};
+
+
+// --------------------------------------------------------------
+// --------------------Base class for testing ------------------------
+// --------------------------------------------------------------
+class TestBase {
+public:
+    MockComMod com_mod;
+    MockCepMod cep_mod;
+};
+
+
+#endif
